+++ editor/vox/vox_editor_plugin.cpp
@@ -1,11 +1,11 @@
 #include "vox_editor_plugin.h"
-#include "vox_importer.h"
 #include "vox_mesh_importer.h"
+#include "vox_scene_importer.h"
 
 namespace zylann::voxel::magica {
 
 VoxEditorPlugin::VoxEditorPlugin(EditorNode *p_node) {
-	Ref<VoxelVoxImporter> vox_scene_importer;
+	Ref<VoxelVoxSceneImporter> vox_scene_importer;
 	vox_scene_importer.instantiate();
 	//add_import_plugin(vox_importer);
 	ResourceFormatImporter::get_singleton()->add_importer(vox_scene_importer);